requirements.txt
Flask==3.0.0
cryptography==42.0.5
argon2-cffi==23.1.0
pycryptodome==3.20.0
Pillow==10.3.0
colorama==0.4.6
